CUD-34: add enum to specify crypto type, remove unused imports, raise exception on keyfile existing to protect against overwriting

+++ appcli/crypto/crypto.py
@@ -26,6 +26,10 @@ def create_key() -> bytes:
 
 
 def create_and_save_key(key_file: Path):
+    if key_file.exists():
+        raise FileExistsError(
+            f"Cannot create keyfile at [{key_file.absolute().as_posix()}] as a file already exists at this path."
+        )
     key_file.write_bytes(create_key())
